fix: adjust grid gutters based on baseline
Change the width of grid gutters in different breakpoints based on the
baseline context.

Breaks gutter definitions. The grid gutter definitions in the `settings` map
must now be a name of a space unit instead of a number (or a function returning
a number). E.g `base` instead of `space(base)`.

Affected settings keys: `grid-gutter`, `grid-gutter-l`, `grid-gutter-t`,
`grid-gutter-xt`, `grid-gutter-xxt`.
The above keys are usually defined in the `settings/_s.grid.scss`
partial in the local projects